requirements.txt
fastapi==0.103.0
uvicorn[standard]==0.23.2
gunicorn==21.2.0
pydantic==2.3.0
pydantic-settings==2.0.3
email-validator==2.0.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
sqlalchemy[asyncio]==2.0.20
asyncpg==0.28.0
alembic==1.11.3
psycopg2-binary==2.9.7
python-multipart==0.0.6
tenacity==8.2.3
